+++ runtimeconfig/manager_test.go
@@ -18,6 +18,7 @@ import (
 	"go.uber.org/atomic"
 	"gopkg.in/yaml.v2"
 
+	"github.com/grafana/dskit/flagext"
 	"github.com/grafana/dskit/services"
 )
 
@@ -65,10 +66,10 @@ func newTestOverridesManagerConfig(t *testing.T, i int32) (*atomic.Int32, Config
 		os.Remove(tempFile.Name())
 	})
 
-	// testing NewRuntimeConfigManager with overrides reload config set
+	// testing runtimeconfig Manager with overrides reload config set
 	return config, Config{
 		ReloadPeriod: 5 * time.Second,
-		LoadPath:     tempFile.Name(),
+		LoadPath:     []string{tempFile.Name()},
 		Loader: func(_ io.Reader) (i interface{}, err error) {
 			val := int(config.Load())
 			return val, nil
@@ -76,27 +77,64 @@ func newTestOverridesManagerConfig(t *testing.T, i int32) (*atomic.Int32, Config
 	}
 }
 
+func generateRuntimeFiles(t *testing.T, overrideStrings []string) ([]*os.File, error) {
+	var overrideFiles []*os.File
+
+	t.Cleanup(func() {
+		require.NoError(t, cleanupOverridesFiles(overrideFiles))
+	})
+
+	for count, override := range overrideStrings {
+		pattern := fmt.Sprintf("overrides-file-%d", count)
+		tempFile, err := os.CreateTemp("", pattern)
+		if err != nil {
+			return nil, err
+		}
+		_, err = tempFile.WriteString(override)
+		if err != nil {
+			return nil, err
+		}
+		overrideFiles = append(overrideFiles, tempFile)
+	}
+
+	return overrideFiles, nil
+}
+
+func generateLoadPath(overrideFiles []*os.File) []string {
+	var fileNames []string
+	for _, f := range overrideFiles {
+		fileNames = append(fileNames, f.Name())
+	}
+	return fileNames
+}
+
+func cleanupOverridesFiles(overrideFiles []*os.File) error {
+	for _, f := range overrideFiles {
+		err := f.Close()
+		if err != nil {
+			return err
+		}
+		err = os.Remove(f.Name())
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
 func TestNewOverridesManager(t *testing.T) {
-	tempFile, err := os.CreateTemp("", "test-validation")
-	require.NoError(t, err)
-
-	defer func() {
-		// Clean up
-		require.NoError(t, tempFile.Close())
-		require.NoError(t, os.Remove(tempFile.Name()))
-	}()
-
-	_, err = tempFile.WriteString(`overrides:
+	tempFiles, err := generateRuntimeFiles(t,
+		[]string{`overrides:
   user1:
-    limit2: 150`)
+    limit2: 150`})
 	require.NoError(t, err)
 
 	defaultTestLimits = &TestLimits{Limit1: 100}
 
-	// testing NewRuntimeConfigManager with overrides reload config set
+	// testing runtimeconfig Manager with overrides reload config set
 	overridesManagerConfig := Config{
 		ReloadPeriod: time.Second,
-		LoadPath:     tempFile.Name(),
+		LoadPath:     generateLoadPath(tempFiles),
 		Loader:       testLoadOverrides,
 	}
 
@@ -109,6 +147,107 @@ func TestNewOverridesManager(t *testing.T) {
 
 	// Make sure test limits were loaded.
 	require.NotNil(t, overridesManager.GetConfig())
+	conf := overridesManager.GetConfig().(*testOverrides)
+	require.NotNil(t, conf)
+	require.Equal(t, 150, conf.Overrides["user1"].Limit2)
+}
+
+func TestOverridesManagerMultipleFilesAppend(t *testing.T) {
+	tempFiles, err := generateRuntimeFiles(t,
+		[]string{`overrides:
+  user1:
+    limit1: 101`,
+			`overrides:
+  user1:
+    limit2: 102`,
+			`overrides:
+  user2:
+    limit1: 103`,
+			`overrides:
+  user2:
+    limit2: 104`})
+	require.NoError(t, err)
+
+	// testing runtimeconfig Manager with overrides reload config set
+	overridesManagerConfig := Config{
+		ReloadPeriod: time.Second,
+		LoadPath:     generateLoadPath(tempFiles),
+		Loader:       testLoadOverrides,
+	}
+
+	overridesManager, err := New(overridesManagerConfig, nil, log.NewNopLogger())
+	require.NoError(t, err)
+	require.NoError(t, services.StartAndAwaitRunning(context.Background(), overridesManager))
+
+	// Cleaning up
+	require.NoError(t, services.StopAndAwaitTerminated(context.Background(), overridesManager))
+
+	// Make sure test limits were loaded.
+	require.NotNil(t, overridesManager.GetConfig())
+	conf := overridesManager.GetConfig().(*testOverrides)
+	require.Equal(t, 101, conf.Overrides["user1"].Limit1)
+	require.Equal(t, 102, conf.Overrides["user1"].Limit2)
+	require.Equal(t, 103, conf.Overrides["user2"].Limit1)
+	require.Equal(t, 104, conf.Overrides["user2"].Limit2)
+}
+
+func TestOverridesManagerMultipleFilesWithOverrides(t *testing.T) {
+	tempFiles, err := generateRuntimeFiles(t,
+		[]string{
+			`overrides:
+  user1:
+    limit1: 100`,
+			`overrides:
+  user1:
+    limit1: 1234`})
+	require.NoError(t, err)
+
+	// testing runtimeconfig Manager with overrides reload config set
+	overridesManagerConfig := Config{
+		ReloadPeriod: time.Second,
+		LoadPath:     flagext.StringSliceCSV(generateLoadPath(tempFiles)),
+		Loader:       testLoadOverrides,
+	}
+
+	overridesManager, err := New(overridesManagerConfig, nil, log.NewNopLogger())
+	require.NoError(t, err)
+	require.NoError(t, services.StartAndAwaitRunning(context.Background(), overridesManager))
+
+	// Cleaning up
+	require.NoError(t, services.StopAndAwaitTerminated(context.Background(), overridesManager))
+
+	// Make sure test limits were loaded.
+	require.NotNil(t, overridesManager.GetConfig())
+	conf := overridesManager.GetConfig().(*testOverrides)
+	require.Equal(t, 1234, conf.Overrides["user1"].Limit1)
+}
+
+func TestOverridesManagerMultipleFilesWithEmptyFile(t *testing.T) {
+	tempFiles, err := generateRuntimeFiles(t,
+		[]string{`overrides:
+  user1:
+    limit1: 100`,
+			``})
+	require.NoError(t, err)
+
+	// testing runtimeconfig Manager with overrides reload config set
+	overridesManagerConfig := Config{
+		ReloadPeriod: time.Second,
+		LoadPath:     generateLoadPath(tempFiles),
+		Loader:       testLoadOverrides,
+	}
+
+	overridesManager, err := New(overridesManagerConfig, nil, log.NewNopLogger())
+	require.NoError(t, err)
+	require.NoError(t, services.StartAndAwaitRunning(context.Background(), overridesManager))
+
+	// Cleaning up
+	require.NoError(t, services.StopAndAwaitTerminated(context.Background(), overridesManager))
+
+	// Make sure test limits were loaded.
+	require.NotNil(t, overridesManager.GetConfig())
+	conf := overridesManager.GetConfig().(*testOverrides)
+	require.Equal(t, 100, conf.Overrides["user1"].Limit1)
 }
 
 func TestManager_ListenerWithDefaultLimits(t *testing.T) {
@@ -122,8 +261,9 @@ func TestManager_ListenerWithDefaultLimits(t *testing.T) {
 	}()
 
 	config := []byte(`overrides:
-  user1:
-    limit2: 150`)
+    user1:
+        limit2: 150
+`)
 	err = os.WriteFile(tempFile.Name(), config, 0600)
 	require.NoError(t, err)
 
@@ -132,7 +272,7 @@ func TestManager_ListenerWithDefaultLimits(t *testing.T) {
 	// testing NewRuntimeConfigManager with overrides reload config set
 	overridesManagerConfig := Config{
 		ReloadPeriod: time.Second,
-		LoadPath:     tempFile.Name(),
+		LoadPath:     []string{tempFile.Name()},
 		Loader:       testLoadOverrides,
 	}
 
@@ -157,8 +297,9 @@ func TestManager_ListenerWithDefaultLimits(t *testing.T) {
 
 	// rewrite file
 	config = []byte(`overrides:
-  user2:
-    limit2: 200`)
+    user2:
+        limit2: 200
+`)
 	err = os.WriteFile(tempFile.Name(), config, 0600)
 	require.NoError(t, err)
 
@@ -270,7 +411,7 @@ func TestManager_ShouldFastFailOnInvalidConfigAtStartup(t *testing.T) {
 	// Create the config manager and start it.
 	cfg := Config{
 		ReloadPeriod: time.Second,
-		LoadPath:     tempFile.Name(),
+		LoadPath:     []string{tempFile.Name()},
 		Loader:       testLoadOverrides,
 	}